feat(hooks): rename  InternalNango module for connection handling
- Updated references in pre-connection and post-connection scripts to utilize the new InternalNango module, replacing the previous shared-hook-logic module.
