+++ packages/server/lib/hooks/connection/providers/docusign/post-connection.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from '../../shared-hook-logic';
+import type { InternalNango as Nango } from '../../internal-nango';
 import type { UserInfoResponse, AccountInfo } from './types.js';
 import { isAxiosError } from 'axios';
