+++ packages/server/lib/hooks/connection/providers/hubspot/post-connection.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from '../../shared-hook-logic';
+import type { InternalNango as Nango } from '../../internal-nango';
 import axios from 'axios';
 
 export default async function execute(nango: Nango) {
