+++ packages/server/lib/hooks/connection/providers/jira/post-connection.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import type { InternalNango as Nango } from '../../shared-hook-logic';
+import type { InternalNango as Nango } from '../../internal-nango';
 
 export default async function execute(nango: Nango) {
     const connection = await nango.getConnection();
