+++ packages/server/lib/hooks/connection/providers/workable/post-connection.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from '../../shared-hook-logic';
+import type { InternalNango as Nango } from '../../internal-nango';
 import { isAxiosError } from 'axios';
 import type { WorkableAccountsResponse } from './types.js';
