+++ packages/server/lib/hooks/connection/providers/xero/post-connection.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from '../../shared-hook-logic';
+import type { InternalNango as Nango } from '../../internal-nango';
 import type { OAuth2Credentials } from '@nangohq/types';
 import jwt from 'jsonwebtoken';
 import { isAxiosError } from 'axios';
